portal/image_app/requirements.txt
django==3.2
djangorestframework==3.12.4
django-extensions==3.1.3
django-cors-headers==3.7.0
boto3==1.16.51
celery==4.4.1
redis==3.4.1
psycopg2>=2.7.5,<2.8.0
requests
Pillow